Restructuring into separate files.
CheckpointMark complete
TimestampPolicy applied to UnboundedRabbitMqReader
RecordIdPolicy appllied to UnboundedRabbitMqReader
Reworked ConnectionHandler
